--- a/includes/libraries/teampassclasses/passwordmanager/src/PasswordManager.php
+++ b/includes/libraries/teampassclasses/passwordmanager/src/PasswordManager.php
@@ -62,7 +62,7 @@ class PasswordManager
         // Vérifiez si le mot de passe a été haché avec passwordlib
         if ($this->isPasswordLibHash($hashedPassword)) {
             // Utilisez la vérification de passwordlib ici
-            if ($this->passwordLibVerify($hashedPassword, $plainPassword)) {
+            if ($this->passwordLibVerify($hashedPassword, html_entity_decode($plainPassword))) {
                 // Password is valid, hash it with new system
                 $newHashedPassword = $this->hashPassword($plainPassword);
                 $userInfo['pw'] = $newHashedPassword;