--- a/install/upgrade_run_3.0.0_suggestions.php
+++ b/install/upgrade_run_3.0.0_suggestions.php
@@ -71,21 +71,14 @@ $database = DB_NAME;
 $port = DB_PORT;
 $user = DB_USER;
 
-if (mysqli_connect(
-    $server,
-    $user,
-    $pass,
-    $database,
-    $port
-)
-) {
-    $db_link = mysqli_connect(
+if ($db_link = mysqli_connect(
         $server,
         $user,
         $pass,
         $database,
         $port
-    );
+    )
+) {
 	$db_link->set_charset(DB_ENCODING);
 } else {
     $res = "Impossible to get connected to server. Error is: ".addslashes(mysqli_connect_error());