--- a/install/upgrade_run_3.0.0_suggestions.php
+++ b/install/upgrade_run_3.0.0_suggestions.php
@@ -160,7 +160,7 @@ while ($data = mysqli_fetch_array($rows)) {
         );
 
         // Encrypt with Object Key
-        $cryptedStuff = doDataEncryption($passwd['string']);
+        $cryptedStuff = doDataEncryption(html_entity_decode($passwd['string']));
 
         // Store new password in DB
         mysqli_query(