**Fix db_link initialization :**
![image](https://github.com/user-attachments/assets/d538abfc-37a1-4477-a709-f35563b7842f)

**Improve scheduler performance :** see issue #4210

**Improve scheduler performance on sharekeys_items generation:** Reduce CPU load keys and generation time.

**Fix TP2 -> TP3 password encoding:**
TP2 :
![Capture d'écran 2024-07-16 114710](https://github.com/user-attachments/assets/89b13162-5e40-4bfc-a245-b32d173b4308)
Json sent to frontend :
![Capture d'écran 2024-07-16 114655](https://github.com/user-attachments/assets/30bc9783-62f5-4da5-b0ae-7c08aea00959)

TP3 :
![Capture d'écran 2024-07-16 115027](https://github.com/user-attachments/assets/690ce340-0a2d-44ab-a3fb-26cbab828ba3)
Same json :
![Capture d'écran 2024-07-16 115007](https://github.com/user-attachments/assets/bcea2932-11fe-4120-88f2-523d045f66ab)

The migration process must absolutely change the password storage strategy.

**Do not sanitize one time codes which may contain & (avoid amp;):**
This is a key generated during upgrade: r&igVj&8kMfOQ&k1
With FILTER_SANITIZE_FULL_SPECIAL_CHARS, '&' becomes '&amp;amp;' and code check fails:
![image](https://github.com/user-attachments/assets/1a79f5fa-94f4-4921-ac4d-f4150fceba21)

**Fix user password migration v2 to v3:**
Users with special characters (&, €, ...) on their personal passwords cannot log back into teampass after upgrade.
We must undo HTML encoding in the migration process.

**Improve v2 -> v3 migration times.**
About 85% time saving during a TP2 to TP3 migration:
- MySQL queries improvements (about 30% time saved):
  - Limit mysql queries in large loops.
  - Use bulk inserts or transactions (avoid autocommit) to speed up.
  - Select count(PK) instead of reporting unnecessary data.
- Increase the number of items processed at a time to reduce the number of requests and avoid browser OutOfMemory (between 15 and 20% time saved - need less than 1 second to process 1000 items).
- Use PHP sessions to keep key items and not decrypt them in each loop (11.5s/1000 items to 1.2s/1000 items -> more than 70% of global time saved).

With my smallest teampass instance of around 5000 items and 30 users, I went from 42 minutes and x seconds of migration to 5 minutes 43 seconds.

**Add composite index on sharekeys_items for big instances.**
For a very big teampass instance (about 600 users / 27 000 items), we need an additional index.
Before:
```sql
mysql> SELECT COUNT(items.id) FROM teampass_items AS items LEFT JOIN teampass_sharekeys_items AS sharekeys_items ON items.id = sharekeys_items.object_id AND sharekeys_items.user_id = 1 WHERE items.perso = 0 LIMIT 0, 30000;
+-----------------+
| COUNT(items.id) |
+-----------------+
|           20141 |
+-----------------+
1 row in set (42,65 sec)

mysql> explain SELECT items.id, items.pw, items.encryption_type, sharekeys_items.share_key FROM teampass_items AS items LEFT JOIN teampass_sharekeys_items AS sharekeys_items ON items.id = sharekeys_items.object_id AND sharekeys_items.user_id = 1 WHERE items.perso = 0 LIMIT 0, 30000;
+----+-------------+-----------------+------------+------+---------------+--------+---------+---------------------+-------+----------+-------------+
| id | select_type | table           | partitions | type | possible_keys | key    | key_len | ref                 | rows  | filtered | Extra       |
+----+-------------+-----------------+------------+------+---------------+--------+---------+---------------------+-------+----------+-------------+
|  1 | SIMPLE      | items           | NULL       | ALL  | NULL          | NULL   | NULL    | NULL                | 25459 |    10.00 | Using where |
|  1 | SIMPLE      | sharekeys_items | NULL       | ref  | OBJECT,USER   | OBJECT | 4       | teampassdb.items.id |   205 |   100.00 | Using where |
+----+-------------+-----------------+------------+------+---------------+--------+---------+---------------------+-------+----------+-------------+
2 rows in set, 1 warning (0,00 sec)
```
After:
```sql
mysql> CREATE INDEX idx_object_user ON teampass_sharekeys_items(object_id, user_id);
Query OK, 0 rows affected (31,65 sec)
Records: 0  Duplicates: 0  Warnings: 0

mysql> explain SELECT items.id, items.pw, items.encryption_type, sharekeys_items.share_key FROM teampass_items AS items LEFT JOIN teampass_sharekeys_items AS sharekeys_items ON items.id = sharekeys_items.object_id AND sharekeys_items.user_id = 1 WHERE items.perso = 0 LIMIT 0, 30000;
+----+-------------+-----------------+------------+------+-----------------------------+-----------------+---------+---------------------------+-------+----------+-------------+
| id | select_type | table           | partitions | type | possible_keys               | key             | key_len | ref                       | rows  | filtered | Extra       |
+----+-------------+-----------------+------------+------+-----------------------------+-----------------+---------+---------------------------+-------+----------+-------------+
|  1 | SIMPLE      | items           | NULL       | ALL  | NULL                        | NULL            | NULL    | NULL                      | 25459 |    10.00 | Using where |
|  1 | SIMPLE      | sharekeys_items | NULL       | ref  | OBJECT,USER,idx_object_user | idx_object_user | 8       | teampassdb.items.id,const |     1 |   100.00 | NULL        |
+----+-------------+-----------------+------------+------+-----------------------------+-----------------+---------+---------------------------+-------+----------+-------------+
2 rows in set, 1 warning (0,01 sec)

mysql> SELECT COUNT(items.id) FROM teampass_items AS items LEFT JOIN teampass_sharekeys_items AS sharekeys_items ON items.id = sharekeys_items.object_id AND sharekeys_items.user_id = 1 WHERE items.perso = 0 LIMIT 0, 30000;
+-----------------+
| COUNT(items.id) |
+-----------------+
|           20141 |
+-----------------+
1 row in set (0,11 sec)
```

**Add index to improve create personal folder scheduler:**
Personal folder scheduled hourly, index added betwen 10h and 11h:
![image](https://github.com/user-attachments/assets/f1777cb9-e028-40a7-8ad5-705f02c7f432)

Before:
```sql
 mysql> explain SELECT id
    ->         FROM teampass_items AS i
    ->         INNER JOIN teampass_log_items AS li ON li.id_item = i.id
    ->         WHERE i.perso = 1 AND li.action = "at_creation" AND li.id_user IN (10000063, 9999997);
+----+-------------+-------+------------+------+--------------------------------+--------------------------------+---------+-----------------+-------+----------+-------------+
| id | select_type | table | partitions | type | possible_keys                  | key                            | key_len | ref             | rows  | filtered | Extra       |
+----+-------------+-------+------------+------+--------------------------------+--------------------------------+---------+-----------------+-------+----------+-------------+
|  1 | SIMPLE      | i     | NULL       | ALL  | PRIMARY                        | NULL                           | NULL    | NULL            | 24394 |    10.00 | Using where |
|  1 | SIMPLE      | li    | NULL       | ref  | teampass_log_items_id_item_IDX | teampass_log_items_id_item_IDX | 4       | teampassdb.i.id |    73 |     2.00 | Using where |
+----+-------------+-------+------------+------+--------------------------------+--------------------------------+---------+-----------------+-------+----------+-------------+
2 rows in set, 1 warning (0,00 sec) 
```
After:
```sql
mysql> explain SELECT id
    ->         FROM teampass_items AS i
    ->         INNER JOIN teampass_log_items AS li ON li.id_item = i.id
    ->         WHERE i.perso = 1 AND li.action = "at_creation" AND li.id_user IN (10000281, 9999997);
+----+-------------+-------+------------+------+---------------------------------------------------------------+--------------------------------+---------+-----------------------+------+----------+--------------------------+
| id | select_type | table | partitions | type | possible_keys                                                 | key                            | key_len | ref                   | rows | filtered | Extra                    |
+----+-------------+-------+------------+------+---------------------------------------------------------------+--------------------------------+---------+-----------------------+------+----------+--------------------------+
|  1 | SIMPLE      | i     | NULL       | ref  | PRIMARY,items_perso_id_idx                                    | items_perso_id_idx             | 1       | const                 | 7180 |   100.00 | Using index              |
|  1 | SIMPLE      | li    | NULL       | ref  | teampass_log_items_id_item_IDX,log_items_item_action_user_idx | log_items_item_action_user_idx | 757     | teampassdb.i.id,const |   11 |    20.00 | Using where; Using index |
+----+-------------+-------+------------+------+---------------------------------------------------------------+--------------------------------+---------+-----------------------+------+----------+--------------------------+
2 rows in set, 1 warning (0,00 sec)
```

Let each request from 0.5s to 0s.